requirements.txt
argparse
discord
ffmpeg-python
pyotp
regex
selenium
SpeechRecognition
urllib3
webdriver_manager
# python-dotenv